+++ FRAuth/FRAuth/Constants/OpenAM.swift
@@ -2,7 +2,7 @@
 //  OpenAM.swift
 //  FRAuth
 //
-//  Copyright (c) 2019-2020 ForgeRock. All rights reserved.
+//  Copyright (c) 2019-2024 ForgeRock. All rights reserved.
 //
 //  This software may be modified and distributed under the terms
 //  of the MIT license. See the LICENSE file for details.
@@ -30,4 +30,6 @@ struct OpenAM {
     static let xRequestedWith = "X-Requested-With"
     static let xmlHTTPRequest = "XMLHTTPRequest"
     static let suspendedId = "suspendedId"
+    static let successUrl = "successUrl"
+    static let realm = "realm"
 }