+++ FRAuth/FRAuth/User/FRUser.swift
@@ -8,7 +8,7 @@
 //  of the MIT license. See the LICENSE file for details.
 //
 
-import Foundation
+import UIKit
 import FRCore