+++ FRAuth/FRAuthTests/FRAuthSwiftTests/E2ETests/Callback-Live/AA-06-DeviceSigningVerifierCallbackTest.swift
@@ -219,10 +219,7 @@ class AA_06_DeviceSigningVerifierCallbackTest: CallbackBaseTest {
                     ex.fulfill()
                 })
                 waitForExpectations(timeout: 60, handler: nil)
-                if isSimulator {
-                    XCTAssertEqual(singningResult, DeviceBindingStatus.clientNotRegistered.errorMessage)
-                    return
-                }
+                
                 XCTAssertEqual(singningResult, "Success")
             }
             else {
@@ -294,10 +291,7 @@ class AA_06_DeviceSigningVerifierCallbackTest: CallbackBaseTest {
                     ex.fulfill()
                 })
                 waitForExpectations(timeout: 60, handler: nil)
-                if isSimulator {
-                    XCTAssertEqual(singningResult, DeviceBindingStatus.clientNotRegistered.errorMessage)
-                    return
-                }
+                
                 XCTAssertEqual(singningResult, "Success")
             }
             else {
@@ -370,10 +364,7 @@ class AA_06_DeviceSigningVerifierCallbackTest: CallbackBaseTest {
                     ex.fulfill()
                 })
                 waitForExpectations(timeout: 60, handler: nil)
-                if isSimulator {
-                    XCTAssertEqual(singningResult, DeviceBindingStatus.clientNotRegistered.errorMessage)
-                    return
-                }
+                
                 XCTAssertEqual(singningResult, "Authentication Timeout")
             }
             else {
@@ -719,10 +710,7 @@ class AA_06_DeviceSigningVerifierCallbackTest: CallbackBaseTest {
                         ex.fulfill()
                     })
                 waitForExpectations(timeout: 60, handler: nil)
-                if isSimulator {
-                    XCTAssertEqual(bindingResult, "DeviceBinding/Signing is not supported on the iOS Simulator")
-                    return
-                }
+                
                 XCTAssertEqual(bindingResult, "Success")
             }
             else {