+++ FRCore/FRCoreTests/FRCore/Network/RestClientTests.swift
@@ -85,7 +85,7 @@ class RestClientTests: FRBaseTestCase {
         XCTAssertNil(error)
     }
     
-    
+    /* Commenting out as httpbin sems flacky at the moment and httpbun does not have an image endpoint
     func test_03_test_invalid_response_image() {
            
        let request = Request(url: FRTestURL.imageURL, method: .GET, headers: ["accept":"image/webp"])
@@ -113,7 +113,7 @@ class RestClientTests: FRBaseTestCase {
        XCTAssertNil(urlResponse)
        
        guard let networkError = error else {
-           XCTFail("Erorr was returned as nil while expecting invalid response error")
+           XCTFail("Error was returned as nil while expecting invalid response error")
            return
        }
        
@@ -125,7 +125,7 @@ class RestClientTests: FRBaseTestCase {
            break
        }
     }
-    
+    */
     
     func test_04_test_cache_control() {