+++ FRGoogleSignIn/FRGoogleSignInTests/GoogleSignInHandlerTests.swift
@@ -2,7 +2,7 @@
 //  GoogleSignInHandlerTests.swift
 //  FRGoogleSignInTests
 //
-//  Copyright (c) 2021 ForgeRock. All rights reserved.
+//  Copyright (c) 2021-2024 ForgeRock. All rights reserved.
 //
 //  This software may be modified and distributed under the terms
 //  of the MIT license. See the LICENSE file for details.
@@ -93,7 +93,7 @@ class GoogleSignInHandlerTests: XCTestCase {
             let googleHandler4 = callback.getDefaultIdPHandler(provider: "google-ios")
             let googleHandler5 = callback.getDefaultIdPHandler(provider: "iosgoogleclient")
             
-            XCTAssertNil(facebookHandler)
+            XCTAssertNotNil(facebookHandler)
             XCTAssertNotNil(appleHandler)
             XCTAssertNotNil(googleHandler1)
             XCTAssertNotNil(googleHandler2)