+++ PingProtect/PingProtectTests/PingOneProtectEvaluationCallbackTests.swift
@@ -218,4 +218,243 @@ final class PingOneProtectEvaluationCallbackTests: FRAuthBaseTest {
         }
     }
     
+    
+    func test_08_derived_callback_init() {
+        let metaDataJsonString = """
+            {
+                "type": "MetadataCallback",
+                "output": [
+                    {
+                        "name": "data",
+                        "value": {
+                            "_type": "PingOneProtect",
+                            "_action": "protect_risk_evaluation",
+                            "envId" : "some_id",
+                            "pauseBehavioralData" : true
+                         }
+                    }
+                ]
+            }
+            """
+        
+        let callbackResponse = self.parseStringToDictionary(metaDataJsonString)
+        
+        do {
+            let callback = try PingOneProtectEvaluationCallback(json: callbackResponse)
+            XCTAssertNotNil(callback)
+        }
+        catch {
+            XCTFail("Failed to construct callback: \(callbackResponse)")
+        }
+    }
+    
+    
+    func test_09_derived_callback_getSignals_success_with_initialization() {
+
+        let jsonStr = """
+        {
+            "authId": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJhdXRoSW5kZXhWYWx1ZSI6IkV4YW1wbGUiLCJvdGsiOiJjOWtvNXQ5Y3JncmdrNTI3MzUxN2RyMXI3YyIsImF1dGhJbmRleFR5cGUiOiJzZXJ2aWNlIiwicmVhbG0iOiIvIiwic2Vzc2lvbklkIjoiKkFBSlRTUUFDTURFQUJIUjVjR1VBQ0VwWFZGOUJWVlJJQUFKVE1RQUEqZXlKMGVYQWlPaUpLVjFRaUxDSmpkSGtpT2lKS1YxUWlMQ0poYkdjaU9pSklVekkxTmlKOS5aWGxLTUdWWVFXbFBhVXBMVmpGUmFVeERTalpoV0VGcFQybEtUMVF3TlVaSmFYZHBXbGMxYWtscWIybFJWRVY1VDBWT1ExRjVNVWxWZWtreFRtbEpjMGx0Um5OYWVVazJTVzFTY0dOcFNqa3VMbGh1VFVSVVJtVlhRMVU0TkRaMllXdG9lRmxoUTFFdVZXRktWazA1ZUVSQ01sQkViRTgwWldwWmJESkVZMXBPWnkxWGJtWnJVelU1UW1SVlQxQlZTbk5tVTB4UFkxZEpWME5HVEdKQ1YzWlNXVmhaTFVkQkxYUmtkM1V3TUcxeFdXVkVlbU0wWlZkcFdXdEtTMmRSV1VoQlpucEZhRVZTU0U4MGJGWkxZalpVVEVnMVJ6VXpTV3RxU0hKQ2EwMUNOMjV0Y25KWk16QlBOMVZLTkVwWlgwWnNjR0pOWldwSlVWUkxhRnBwZFVwaVJsaHdOMDVqYXpOU1FVdGpaMW96VjNScFdpMXdZM1YzUzJkM2NIVkxVRnB2UTJRMFJYVkRRakJmUTBnNGMxODJUMnR1ZEMxM1dVcGtkVEZKU2xobU5UWTBjakowYkd4UmJuaEVZbGRqYlV0V01taFFTVmRKZEZaTE1XSlNPRFI0ZUhNMFVGQkRiblV5VkU0dE1WRnNUazlwWmpsYVJUa3pNV3RETTE5MVRGQnZXVFF0V1VsekxWaFdOVzVUWlRab1F6STJhalUyWlZSbFRYQktja1U0TWpWbGNHZ3pPRXBSVDNCT1VtMTJkVGhQV1ZCTmEwcHdNQzB3VkUxMFIxbFdOamhoWTB0RlkxVjBibmxyVGxWVk4wVkxSa2N0TjFnNFl6VnFSMVV3U2w5YWFUTkZhMk0wT1hWdFFubFRiMms0UWtwTFdtaFdZak50VUZWMlpUSlhlSFZzUzJObVZuTXpTbHBOTlVJek1tTlVTbDloWldsNk9HOXRjRTFwUTB0dE1UUXllalZFTUZWS2FESk9YelkxVFZKR1ZEWnVXSE5mY2w5YWVrRk1aRFZKZERWemJIUkxTVUZEYlVsc01rOXljVlF3ZDNodllXWnFTWGx3V2tNelJEZHZSMmxwY25SeFpuaElNRmxRWDFwVU9FbDRkbHBTVlcxYWFHSlFhVzQwU1hWMFQxUTRPV2wxZEdwd2RYUlpWRmhXTkZoV2NtdFlRVWgyV0hVeFNGTXdWbXh0VG5NNVNWOU5SMjlpVTBoeGVYazBTalY1ZGs1M1pqTklkeTQ1V2pBeVZYazNSMDVUY2xsNGJqaG1Nbk10ZUZabi43SHNDVlNMUUljREk5S1pzQ1N0cjJEM3BTQmJhV1A1UlY2T29pX0lnODA0IiwiZXhwIjoxNTYyNzg4MDAyLCJpYXQiOjE1NjI3ODc3MDJ9.oEiBLxT62uwmz0EtLQxwzjyrgcIy7fpevO6TntEK8aM",
+            "callbacks": [
+                {
+                    "type": "MetadataCallback",
+                    "output": [
+                        {
+                            "name": "data",
+                            "value": {
+                                "_type": "PingOneProtect",
+                                "_action": "protect_risk_evaluation",
+                                "envId" : "some_id",
+                                "pauseBehavioralData" : true
+                             }
+                        }
+                    ]
+                },
+                {
+                    "type": "HiddenValueCallback",
+                    "output": [
+                        {
+                            "name": "value",
+                            "value": ""
+                        },
+                        {
+                            "name": "id",
+                            "value": "pingone_risk_evaluation_signals"
+                        }
+                    ],
+                    "input": [
+                        {
+                            "name": "IDToken1",
+                            "value": "pingone_risk_evaluation_signals"
+                        }
+                    ]
+                },
+                {
+                    "type": "HiddenValueCallback",
+                    "output": [
+                        {
+                            "name": "value",
+                            "value": ""
+                        },
+                        {
+                            "name": "id",
+                            "value": "clientError"
+                        }
+                    ],
+                    "input": [
+                        {
+                            "name": "IDToken1",
+                            "value": "clientError"
+                        }
+                    ]
+                }
+            ]
+        }
+        """
+        let authServiceResponse = self.parseStringToDictionary(jsonStr)
+        let serverConfig = ServerConfigBuilder(url: URL(string: "http://localhost:8080/am")!, realm: "customRealm").set(timeout: 90.0).build()
+        
+        CallbackFactory.shared.registerCallback(callbackType: ProtectCallbackType.riskEvaluation.rawValue, callbackClass: PingOneProtectEvaluationCallback.self)
+        
+        do {
+            let node = try Node(UUID().uuidString, authServiceResponse, serverConfig, "serviceName", "service")
+            XCTAssertNotNil(node)
+            
+            // Expect PingOneProtectEvaluationCallback callback
+            for callback in node!.callbacks {
+                if callback is PingOneProtectEvaluationCallback, let pingOneProtectEvaluationCallback = callback as? PingOneProtectEvaluationCallback {
+                    
+                    XCTAssertTrue(pingOneProtectEvaluationCallback.pauseBehavioralData)
+                    
+                    let ex1 = self.expectation(description: "SDK initialized")
+                    PIProtect.start { error in
+                        XCTAssertNil(error)
+                        ex1.fulfill()
+                    }
+                    
+                    var evaulationResult = ""
+                    let ex2 = self.expectation(description: "PingOne Protect Evaluate")
+                    pingOneProtectEvaluationCallback.getData(completion: { (result) in
+                            switch result {
+                            case .success:
+                                evaulationResult = "Success"
+                            case .failure(let error):
+                                evaulationResult = error.localizedDescription
+                            };
+                            ex2.fulfill()
+                        })
+                    waitForExpectations(timeout: 5, handler: nil)
+                    
+                    XCTAssertEqual(evaulationResult, "Success")
+                    
+                    // Ensure that Signals data is not empty after collection
+                    var signals: String?
+                    for callback in node!.callbacks {
+                        if callback is HiddenValueCallback, let hiddenCallback = callback as? HiddenValueCallback, let callbackId = hiddenCallback.id, callbackId.contains(CBConstants.riskEvaluationSignals) {
+                            signals = hiddenCallback.getValue() as? String
+                        }
+                    }
+                    XCTAssertNotNil(signals)
+                }
+            }
+        }
+        catch {
+            XCTFail("Failed to construct node: \(jsonStr)")
+        }
+    }
+    
+    
+    func test_10_derived_callback_client_error() {
+        
+        let jsonStr = """
+        {
+            "authId": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJhdXRoSW5kZXhWYWx1ZSI6IkV4YW1wbGUiLCJvdGsiOiJjOWtvNXQ5Y3JncmdrNTI3MzUxN2RyMXI3YyIsImF1dGhJbmRleFR5cGUiOiJzZXJ2aWNlIiwicmVhbG0iOiIvIiwic2Vzc2lvbklkIjoiKkFBSlRTUUFDTURFQUJIUjVjR1VBQ0VwWFZGOUJWVlJJQUFKVE1RQUEqZXlKMGVYQWlPaUpLVjFRaUxDSmpkSGtpT2lKS1YxUWlMQ0poYkdjaU9pSklVekkxTmlKOS5aWGxLTUdWWVFXbFBhVXBMVmpGUmFVeERTalpoV0VGcFQybEtUMVF3TlVaSmFYZHBXbGMxYWtscWIybFJWRVY1VDBWT1ExRjVNVWxWZWtreFRtbEpjMGx0Um5OYWVVazJTVzFTY0dOcFNqa3VMbGh1VFVSVVJtVlhRMVU0TkRaMllXdG9lRmxoUTFFdVZXRktWazA1ZUVSQ01sQkViRTgwWldwWmJESkVZMXBPWnkxWGJtWnJVelU1UW1SVlQxQlZTbk5tVTB4UFkxZEpWME5HVEdKQ1YzWlNXVmhaTFVkQkxYUmtkM1V3TUcxeFdXVkVlbU0wWlZkcFdXdEtTMmRSV1VoQlpucEZhRVZTU0U4MGJGWkxZalpVVEVnMVJ6VXpTV3RxU0hKQ2EwMUNOMjV0Y25KWk16QlBOMVZLTkVwWlgwWnNjR0pOWldwSlVWUkxhRnBwZFVwaVJsaHdOMDVqYXpOU1FVdGpaMW96VjNScFdpMXdZM1YzUzJkM2NIVkxVRnB2UTJRMFJYVkRRakJmUTBnNGMxODJUMnR1ZEMxM1dVcGtkVEZKU2xobU5UWTBjakowYkd4UmJuaEVZbGRqYlV0V01taFFTVmRKZEZaTE1XSlNPRFI0ZUhNMFVGQkRiblV5VkU0dE1WRnNUazlwWmpsYVJUa3pNV3RETTE5MVRGQnZXVFF0V1VsekxWaFdOVzVUWlRab1F6STJhalUyWlZSbFRYQktja1U0TWpWbGNHZ3pPRXBSVDNCT1VtMTJkVGhQV1ZCTmEwcHdNQzB3VkUxMFIxbFdOamhoWTB0RlkxVjBibmxyVGxWVk4wVkxSa2N0TjFnNFl6VnFSMVV3U2w5YWFUTkZhMk0wT1hWdFFubFRiMms0UWtwTFdtaFdZak50VUZWMlpUSlhlSFZzUzJObVZuTXpTbHBOTlVJek1tTlVTbDloWldsNk9HOXRjRTFwUTB0dE1UUXllalZFTUZWS2FESk9YelkxVFZKR1ZEWnVXSE5mY2w5YWVrRk1aRFZKZERWemJIUkxTVUZEYlVsc01rOXljVlF3ZDNodllXWnFTWGx3V2tNelJEZHZSMmxwY25SeFpuaElNRmxRWDFwVU9FbDRkbHBTVlcxYWFHSlFhVzQwU1hWMFQxUTRPV2wxZEdwd2RYUlpWRmhXTkZoV2NtdFlRVWgyV0hVeFNGTXdWbXh0VG5NNVNWOU5SMjlpVTBoeGVYazBTalY1ZGs1M1pqTklkeTQ1V2pBeVZYazNSMDVUY2xsNGJqaG1Nbk10ZUZabi43SHNDVlNMUUljREk5S1pzQ1N0cjJEM3BTQmJhV1A1UlY2T29pX0lnODA0IiwiZXhwIjoxNTYyNzg4MDAyLCJpYXQiOjE1NjI3ODc3MDJ9.oEiBLxT62uwmz0EtLQxwzjyrgcIy7fpevO6TntEK8aM",
+            "callbacks": [
+                {
+                    "type": "MetadataCallback",
+                    "output": [
+                        {
+                            "name": "data",
+                            "value": {
+                                "_type": "PingOneProtect",
+                                "_action": "protect_risk_evaluation",
+                                "envId" : "some_id",
+                                "pauseBehavioralData" : true
+                             }
+                        }
+                    ]
+                },
+                {
+                    "type": "HiddenValueCallback",
+                    "output": [
+                        {
+                            "name": "value",
+                            "value": ""
+                        },
+                        {
+                            "name": "id",
+                            "value": "pingone_risk_evaluation_signals"
+                        }
+                    ],
+                    "input": [
+                        {
+                            "name": "IDToken1",
+                            "value": "pingone_risk_evaluation_signals"
+                        }
+                    ]
+                },
+                {
+                    "type": "HiddenValueCallback",
+                    "output": [
+                        {
+                            "name": "value",
+                            "value": ""
+                        },
+                        {
+                            "name": "id",
+                            "value": "clientError"
+                        }
+                    ],
+                    "input": [
+                        {
+                            "name": "IDToken1",
+                            "value": "clientError"
+                        }
+                    ]
+                }
+            ]
+        }
+        """
+        let authServiceResponse = self.parseStringToDictionary(jsonStr)
+        let serverConfig = ServerConfigBuilder(url: URL(string: "http://localhost:8080/am")!, realm: "customRealm").set(timeout: 90.0).build()
+        
+        CallbackFactory.shared.registerCallback(callbackType: ProtectCallbackType.riskEvaluation.rawValue, callbackClass: PingOneProtectEvaluationCallback.self)
+        
+        do {
+            
+            let node = try Node(UUID().uuidString, authServiceResponse, serverConfig, "serviceName", "service")
+            XCTAssertNotNil(node)
+            
+            // Expect PingOneProtectEvaluationCallback callback
+            var pingOneProtectEvaluationCallback: PingOneProtectEvaluationCallback?
+            for callback in node!.callbacks {
+                if callback is PingOneProtectEvaluationCallback {
+                    pingOneProtectEvaluationCallback = callback as? PingOneProtectEvaluationCallback
+                    pingOneProtectEvaluationCallback?.setClientError("Some failure!")
+                }
+            }
+            XCTAssertNotNil(pingOneProtectEvaluationCallback)
+            
+            var clientError: String?
+            for callback in node!.callbacks {
+                if callback is HiddenValueCallback, let hiddenCallback = callback as? HiddenValueCallback, let callbackId = hiddenCallback.id, callbackId.contains("clientError") {
+                    clientError = hiddenCallback.getValue() as? String
+                }
+            }
+            XCTAssertNotNil(clientError)
+            XCTAssertEqual(clientError, "Some failure!")
+        }
+        catch {
+            XCTFail("Failed to construct node: \(jsonStr)")
+        }
+    }
 }